metagam: Meta-analysis of generalized additive models.

The main functions in the metagam package are described below.

Stripping rawdata

The function strip_rawdata takes a fit produced by functions in packages mgcv or gamm4 and removes all individual participants data.

Meta-analysis

The function metagam takes a list of fits produced by strip_rawdata and computes meta-analytic fits.

Plotting

The functions plot_dominance and plot_heterogeneity can be used to study the meta-analytic fit computed by strip_rawdata.

See also

Author

Maintainer: Oystein Sorensen oystein.sorensen@psykologi.uio.no (ORCID)

Authors:

Andreas M. Brandmaier brandmaier@mpib-berlin.mpg.de (ORCID)

Athanasia Mo Mowinckel a.m.mowinckel@psykologi.uio.no (ORCID)